moved BaseMessage to __init__.py
