+++ src/events/urgent_message.py
@@ -1,5 +1,13 @@
+from . import BaseMessage
+
 class IdentifiedUrgentMessage:
-    pass
+    _REQUIRED_ATTR_PLACEHOLDER="placeholder"
+
+    def __init__(self, content: dict):
+        super().__init__(content, required_attrs=[self._REQUIRED_ATTR_PLACEHOLDER])
 
 class RespondUrgentMessage:
-    pass
+    _REQUIRED_ATTR_PLACEHOLDER="placeholder"
+
+    def __init__(self, content: dict):
+        super().__init__(content, required_attrs=[self._REQUIRED_ATTR_PLACEHOLDER])
